명함 사용자 정보 조회(멀티건) 카운트 단위 테스트

+++ src/main/java/egovframework/com/cop/ncm/service/impl/NcrdManageDAO.java
@@ -118,7 +118,7 @@ public class NcrdManageDAO extends EgovComAbstractDAO {
      * @throws Exception
      */
     public int selectNcrdUseInfsCnt(NameCardUser ncrdUser) throws Exception {
-	return (Integer)selectOne("NcrdManageDAO.selectNcrdUseInfsCnt", ncrdUser);
+        return selectOne("NcrdManageDAO.selectNcrdUseInfsCnt", ncrdUser);
     }
 
     /**
